+++ NEWS.md
@@ -14,6 +14,7 @@ R to use this package as R "3.50." It should have been R 3.5.0 and now reflects
 that it should work on versions of R greater than or equal to 3.5.0. (#126)
 * The `effect_plot()` function now accepts the argument `line.colors` to allow
 users to request different colors for the line/interval and plotted points. 
+* Removed the soft dependency on the `ggstance` package.
 
 # jtools 2.2.0